+++ procedural/scheduling/MODULE_DOCS.md
@@ -1,1 +1,21 @@
 # Module Scheduling
+
+This library provides tools for creating scheduling procedures, based on the timeline library.
+To write a procedure, create a class that implements the Procedure interface.
+Then, follow the tutorial documentation (TODO: link to tutorial documentation once written)
+to package and upload your constraint to Aerie.
+
+# Package gov.nasa.jpl.aerie.procedural.scheduling
+The top-level procedure types.
+
+# Package gov.nasa.jpl.aerie.procedural.scheduling.plan
+Contains the EditablePlan interface that procedure execution environments need to
+implement.
+
+# Package gov.nasa.jpl.aerie.procedural.scheduling.simulation
+Contains configuration options for spawning simulations.
+
+# Package gov.nasa.jpl.aerie.procedural.scheduling.annotations
+Annotations for serializing and deserializing procedure arguments.
+
+
