+++ procedural/scheduling/src/main/kotlin/gov/nasa/jpl/aerie/procedural/scheduling/Procedure.kt
@@ -3,6 +3,7 @@ package gov.nasa.jpl.aerie.procedural.scheduling
 import gov.nasa.jpl.aerie.procedural.scheduling.plan.EditablePlan
 import gov.nasa.jpl.aerie.timeline.CollectOptions
 
+/** The interface that all scheduling procedures must satisfy. */
 interface Procedure {
   /**
    * Run the procedure.
