+++ procedural/timeline/src/main/kotlin/gov/nasa/jpl/aerie/timeline/plan/SqlUtils.kt
@@ -3,7 +3,7 @@ package gov.nasa.jpl.aerie.timeline.plan
 import java.sql.Connection
 import java.sql.PreparedStatement
 
-fun getSingleLongQueryResult(statement: PreparedStatement): Long {
+internal fun getSingleLongQueryResult(statement: PreparedStatement): Long {
   val result = statement.executeQuery()
   if (!result.next()) throw AeriePostgresSimulationResults.DatabaseError("Expected exactly one result for query, found none: $statement")
   val int = result.getLong(1)
@@ -11,4 +11,4 @@ fun getSingleLongQueryResult(statement: PreparedStatement): Long {
   return int
 }
 
-fun intervalStyleStatement(c: Connection): PreparedStatement = c.prepareStatement("set intervalstyle = 'iso_8601';")
+internal fun intervalStyleStatement(c: Connection): PreparedStatement = c.prepareStatement("set intervalstyle = 'iso_8601';")
